Programowanie komponentowe/Zadanie11b/Model/src/main/java/pl/cp/first/FileSudokuBoardDao.java
package pl.cp.first;

import java.io.FileInputStream;
import java.io.FileOutputStream;
import java.io.IOException;
import java.io.ObjectInputStream;
import java.io.ObjectOutputStream;
import pl.cp.first.exception.FileIoException;

public class FileSudokuBoardDao implements Dao<SudokuBoard> {

    private final String fileName;

    public FileSudokuBoardDao(String fileName) {
        this.fileName = fileName;
    }

    @Override
    public SudokuBoard read() throws FileIoException {
        SudokuBoard sudokuBoard;
        try (ObjectInputStream objectInputStream =
                     new ObjectInputStream(new FileInputStream(fileName))) {
            sudokuBoard = (SudokuBoard) objectInputStream.readObject();
        } catch (IOException | ClassNotFoundException e) {
            throw new FileIoException(e);
        }
        return sudokuBoard;
    }

    @Override
    public void write(SudokuBoard sudokuBoard) throws FileIoException {
        try (ObjectOutputStream outputStream =
                     new ObjectOutputStream(new FileOutputStream(fileName))) {
            outputStream.writeObject(sudokuBoard);
        } catch (IOException e) {
            throw new FileIoException(e);
        }
    }

    @Override
    public void close() {
    }
}